chore: add banner types
